@@ -5,12 +5,19 @@ class PasswordInput extends StatefulWidget {
   final Function validator;
   final Function onChanged;
   final TextEditingController controller;
+  final bool showBtn;
+  final TextStyle textStyle;
+  final bool autofocus;
+
 
   PasswordInput(
       {@required this.label,
       @required this.validator,
       @required this.onChanged,
-      @required this.controller});
+      @required this.controller,
+      this.showBtn = true,
+      this.textStyle,
+      this.autofocus = false,});
 
   @override
   _PasswordInputState createState() => _PasswordInputState();
@@ -22,21 +29,23 @@ class _PasswordInputState extends State<PasswordInput> {
   @override
   Widget build(BuildContext context) {
     return TextFormField(
+      autofocus: widget.autofocus,
       // onFieldSubmitted: ,
       onChanged: (String v) {
         widget.onChanged(v);
       },
       obscureText: !show,
       textInputAction: TextInputAction.done,
-      style: TextStyle(fontSize: 16.0),
+      style: widget.textStyle ?? TextStyle(fontSize: 16.0),
       controller: widget.controller,
       validator: widget.validator,
+      cursorColor: widget.textStyle != null ? widget.textStyle.color : Theme.of(context).primaryColor,
       decoration: InputDecoration(
         focusedBorder: OutlineInputBorder(
           borderSide: BorderSide(color: Theme.of(context).primaryColorLight),
         ),
         enabledBorder: OutlineInputBorder(
-          borderSide: BorderSide(color: Theme.of(context).cursorColor),
+          borderSide: BorderSide(color: widget.showBtn ? Theme.of(context).cursorColor : Colors.transparent),
         ),
         contentPadding: EdgeInsets.all(8.0),
         // labelStyle: widget.validator(widget.controller.text) == null
@@ -44,9 +53,8 @@ class _PasswordInputState extends State<PasswordInput> {
         //     : TextStyle(color: Colors.red, fontSize: 14.0),
         labelStyle: TextStyle(color: Theme.of(context).accentColor),
         labelText: widget.label,
-
         // errorStyle: TextStyle(color: Colors.red, fontSize: 10.0),
-        suffixIcon: GestureDetector(
+        suffixIcon: widget.showBtn ? GestureDetector(
             onTap: () {
               setState(() {
                 show = !show;
@@ -54,7 +62,7 @@ class _PasswordInputState extends State<PasswordInput> {
             },
             child: show
                 ? ImageIcon(AssetImage('assets/images/icons/ic_openeye.png'))
-                : ImageIcon(AssetImage('assets/images/icons/ic_closeeye.png'))),
+                : ImageIcon(AssetImage('assets/images/icons/ic_closeeye.png'))) : null,
       ),
     );
   }
